+++ docs/RELEASE_NOTES.md
@@ -42,9 +42,11 @@
 [Javadoc](https://commercetools.github.io/commercetools-sync-java/v/1.5.0/) | 
 [Jar](https://bintray.com/commercetools/maven/commercetools-sync-java/1.5.0)
 
-- 🎉 **New Features** (2)
+- 🎉 **New Features** (4)
     - **Product Sync** - Added support for resolving product references on attributes of type `Reference`, `Set` of `Reference`, `NestedType` or `Set` of `NestedType`. [#438](https://github.com/commercetools/commercetools-sync-java/issues/438)
     - **Product Sync** - Added support for resolving category references on attributes of type `Reference`, `Set` of `Reference`, `NestedType` or `Set` of `NestedType`. [#440](https://github.com/commercetools/commercetools-sync-java/issues/440)
+    - **Product Sync** - Added support for resolving productType references on attributes of type `Reference`, `Set` of `Reference`, `NestedType` or `Set` of `NestedType`. [#440](https://github.com/commercetools/commercetools-sync-java/issues/443)
+    - **Commons** - Exposed `ResourceIdentifierUtils#isReferenceOfType` util which checks if a JSON representation of a CTP `Reference` object is of a certain type or not. [#443](https://github.com/commercetools/commercetools-sync-java/issues/443)
     
 - 🚧 **Breaking Changes** (1)
     - **Product Sync** - Unexposed the methods `VariantReferenceResolver#resolveAttributeReferences` and `VariantReferenceResolver#resolveAttributeReference` to be `private` as they are only meant for internal use of the library. [#440](https://github.com/commercetools/commercetools-sync-java/issues/440)